pyproject.toml
[project]
name = "multi-agent-video-system"
version = "0.1.0"
description = "AI-powered video creation platform built on Google's Agent Development Kit (ADK) framework"
authors = [
    { name = "Video System Team", email = "team@example.com" },
]

license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.8.1"
tabulate = "^0.9.0"
google-auth = "^2.36.0"
requests = "^2.32.3"
google-adk = "^1.7.0"
google-cloud-aiplatform = { extras = [
    "adk",
    "agent-engines",
], version = "^1.93.0" }
llama-index = "^0.12"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
ffmpeg-python = "^0.2.0"
# Error handling and resilience dependencies
psutil = "^6.1.0"
tenacity = "^9.0.0"
# Session management and persistence dependencies
pymongo = "^4.6.0"
motor = "^3.3.0"
# Concurrent processing and performance monitoring dependencies
aiofiles = "^24.1.0"
asyncio-throttle = "^1.0.2"
# CLI and API interface dependencies
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
click = "^8.1.0"
rich = "^13.9.0"
# Real video generation dependencies
moviepy = "^1.0.3"
pillow = "^10.0.0"
pyttsx3 = "^2.90"
numpy = "^1.24.0"
aiohttp = "^3.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
scikit-learn = "^1.6.1"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"